pyproject.toml
[project]
name = "multi-agent-personal-finance-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "google-adk>=0.1.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
